starting the separation
